+++ src/domain/dataCtrl/PostDataCtrl.java
@@ -7,7 +7,7 @@ import domain.Post;
 public interface PostDataCtrl {
     public boolean insert(Post p);
     public List<Post> selectAll();
-    public List<Post> selectByDateBigger(Integer firstDate);
-    public List<Post> selectByDateSmaller(Integer lastDate);
-    public List<Post> selectByDateBoth(Integer firstDate, Integer lastDate);
+    public List<Post> selectByDateBigger(Long firstDate);
+    public List<Post> selectByDateSmaller(Long lastDate);
+    public List<Post> selectByDateBoth(Long firstDate, Long lastDate);
 }
